requirements.txt
langchain
langgraph
langchain-openai
pypdf
pydantic
SpeechRecognition
gTTS
python-dotenv
openai